+++ test/core/UtilTest.cpp
@@ -35,16 +35,16 @@ class UtilTest : public QObject {
     for (int i = 0; i < 100; i++) {
       vec.append(i);
     }
-    size_t idx = Util::binarySearch(vec.length(), [vec](size_t i) { return i > 70; });
-    QCOMPARE(static_cast<int>(idx), 70);
+    int idx = Util::binarySearch(vec.length(), [vec](int i) { return i > 70; });
+    QCOMPARE(idx, 70);
 
     // binarySearch returns vec.length if there's no element which returns f(i) == true
-    idx = Util::binarySearch(vec.length(), [vec](size_t) { return false; });
-    QCOMPARE(static_cast<int>(idx), vec.length());
+    idx = Util::binarySearch(vec.length(), [vec](int) { return false; });
+    QCOMPARE(idx, vec.length());
 
     // binarySearch returns 0 if all the items satisfies the predicate.
-    idx = Util::binarySearch(vec.length(), [vec](size_t i) { return i <= 100; });
-    QCOMPARE(static_cast<int>(idx), 0);
+    idx = Util::binarySearch(vec.length(), [vec](int i) { return i <= 100; });
+    QCOMPARE(idx, 0);
   }
 
   void toStdStringList() {
